Agregado de clase Polideportivo junto con métodos de interfaces

+++ src/ejercicio2/interfazEdificio.java
@@ -2,5 +2,5 @@ package ejercicio2;
 
 public interface interfazEdificio {
 
-	int getSuperficieEdificio();
+	double getSuperficieEdificio();
 }
